Moved HTML, CSS, and JavaScript files to main directory for GitHub Pages
GitHub Pages was unable to locate the index.html file because it was nested inside specific folders. To resolve this issue and enable GitHub Pages to serve the website properly, I moved all the necessary files (HTML, CSS, and JavaScript) to the main directory.

This change ensures that the website can be accessed correctly when deployed using GitHub Pages.